server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
google-adk==1.0.0
google-cloud-firestore==2.20.0
google-cloud-tasks==2.17.0
python-dotenv==1.0.1
httpx==0.28.1
beautifulsoup4==4.12.3
requests==2.32.3
